+++ Cargo.toml
@@ -25,6 +25,9 @@ derive_more = { version = "1.0.0", features = ["full"] }
 dotenv = "0.15.0"
 envy = "0.4.2"
 futures = "0.3.30"
+http = "1.1.0"
+http-body-util = "0.1.2"
+mime = "0.3.17"
 paste = "1.0.15"
 pretty_assertions = "1.4.1"
 reqwest = { version = "0.12.5", features = ["json", "blocking"] }
@@ -33,9 +36,11 @@ serde_json = { version = "1.0.121" }
 sqlx = { version = "0.8.0", features = ["runtime-tokio", "postgres", "json"] }
 thiserror = "1.0.63"
 tokio = { version = "1.39.2", features = ["full"] }
+tower-http = { version = "0.6.1", features = ["trace"] }
 tracing = "0.1.40"
 tracing-subscriber = "0.3.18"
 
 [dev-dependencies]
 insta = "1.39.0"
 tokio = { version = "1.39.2", features = ["full", "test-util"] }
+tower = "0.5.1"
